biofilmVolumeFractionIdx: make unsigned

+++ opm/models/blackoil/blackoilnewtonmethod.hpp
@@ -410,7 +410,7 @@ protected:
                 if (pvIdx == static_cast<int>(Indices::microbialConcentrationIdx)) {
                     nextValue[pvIdx] = std::max(nextValue[pvIdx], Scalar{0.0});
                 }
-                if (pvIdx == Indices::biofilmVolumeFractionIdx) {
+                if (pvIdx == static_cast<int>(Indices::biofilmVolumeFractionIdx)) {
                     nextValue[pvIdx] = std::clamp(nextValue[pvIdx],
                                                   Scalar{0.0},
                                                   this->problem().referencePorosity(globalDofIdx, 0) - 1e-8);
